requirements.txt
Pillow
watchdog
